Restructure campaign dir
