+++ lib/middleware/campaigns/index/contentful-keywords.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const contentful = require('../../contentful');
-const helpers = require('../../helpers');
+const contentful = require('../../../contentful');
+const helpers = require('../../../helpers');
 
 module.exports = function getContentfulKeywords() {
   return (req, res, next) => {
