+++ lib/middleware/campaigns/index/format-campaigns.js
