+++ lib/middleware/campaigns/index/phoenix-campaigns.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const helpers = require('../../helpers');
+const helpers = require('../../../helpers');
 
 module.exports = function getPhoenixCampaigns() {
   return (req, res, next) => {
