+++ lib/middleware/campaigns/single/bot-config-get.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const helpers = require('../../helpers');
+const helpers = require('../../../helpers');
 
 module.exports = function getBotConfig() {
   return (req, res, next) => helpers.botConfig.fetchByCampaignId(req.campaignId)
