+++ lib/middleware/campaigns/single/bot-config-parse.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const helpers = require('../../helpers');
+const helpers = require('../../../helpers');
 
 module.exports = function parseBotConfig() {
   return (req, res) => {
