+++ lib/middleware/campaigns/single/campaign-get.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const helpers = require('../../helpers');
+const helpers = require('../../../helpers');
 
 module.exports = function getCampaign() {
   return (req, res, next) => {
