+++ lib/middleware/campaigns/single/keywords-get.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const contentful = require('../../contentful');
-const helpers = require('../../helpers');
+const contentful = require('../../../contentful');
+const helpers = require('../../../helpers');
 
 module.exports = function getKeywords() {
   return (req, res, next) => {
